requirements.txt
# File: backend/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
asyncpg
sqlalchemy
httpx
openai  # DeepSeek uses OpenAI-compatible API
PyPDF2  # For PDF text extraction
supabase
pydantic
pydantic-settings